Fix WebSocket protocol types, migrate to unified /ws endpoint, update to net8/net10, convert tests to xunit, update CI/CD
Agent-Logs-Url: https://github.com/IoTSharp/TaosConnector/sessions/8e77872f-e71a-4224-9cc6-b665bd79d063

+++ src/IoTSharp.Data.Taos/Protocols/TDWebSocket/WSConnReq.cs
@@ -3,7 +3,7 @@
 
     public class WSConnReq
     {
-        public int req_id { get; set; }
+        public long req_id { get; set; }
         public string user { get; set; }
         public string password { get; set; }
         public string db { get; set; }
